Add offset to gcm
Tweaked encryption/decryption in gcm functions to allow for offset

+++ userspace-module/utils/crypt-utils/crypt-utils.h
@@ -38,7 +38,7 @@
  * */
 #define IV_SIZE 16
 
-extern int do_crypt (int mode, FILE *fp, unsigned char **text, int len,
+extern int do_crypt (int mode, FILE *fp, unsigned char **text, size_t len,
                      unsigned char *key, unsigned char *iv, off_t offset);
 
 void generate_key (unsigned char *destination);
